+++ src/plugins/stake-plugins/generic/policyAdapters/EthereumKilnAdaptor.ts
@@ -22,10 +22,11 @@ export interface EthereumPooledKilnAdapterConfig {
   rpcProviderUrls: string[]
 }
 
-export const makeKilnAdapter = (policyConfig: StakePolicyConfig<EthereumPooledKilnAdapterConfig>): StakePolicyAdapter => {
+export const makeEthereumKilnAdapter = (policyConfig: StakePolicyConfig<EthereumPooledKilnAdapterConfig>): StakePolicyAdapter => {
   const { stakePolicyId, adapterConfig } = policyConfig
   const { apiKey, baseUrl, contractAddress, exitQueueAddress, rpcProviderUrls } = adapterConfig
-  if (apiKey == null) throw new Error(`Kiln apikey is required for ${stakePolicyId}`)
+
+  if (apiKey == null) throw new Error(`Kiln apiKey is required for ${stakePolicyId}`)
 
   const kiln = makeKilnApi(baseUrl, apiKey)
 
@@ -81,7 +82,7 @@ export const makeKilnAdapter = (policyConfig: StakePolicyConfig<EthereumPooledKi
 
       const allocations: QuoteAllocation[] = []
 
-      const operations = await kiln.getOperations(walletAddress)
+      const operations = await kiln.ethGetOnChainOperations(walletAddress)
 
       let claimableTotal = '0'
       const ticketIds: string[] = []
@@ -164,7 +165,7 @@ export const makeKilnAdapter = (policyConfig: StakePolicyConfig<EthereumPooledKi
       let canClaim = false
       const allocations: PositionAllocation[] = []
 
-      const allPositions = await kiln.getStakes(walletAddress)
+      const allPositions = await kiln.ethGetOnChainStakes(walletAddress)
       const position = allPositions.find(position => position.integration_address === contractAddress)
 
       // After fully unstaking, users are left with a single wei of the liquidity token. We should ignore this.
@@ -185,7 +186,7 @@ export const makeKilnAdapter = (policyConfig: StakePolicyConfig<EthereumPooledKi
         nativeAmount: position?.rewards ?? '0'
       })
 
-      const operations = await kiln.getOperations(walletAddress)
+      const operations = await kiln.ethGetOnChainOperations(walletAddress)
 
       let claimableTotal = '0'
       for (const operation of operations) {